+++ packages/x-charts/src/RadarChart/RadarGrid/radarGridClasses.ts
@@ -1,7 +1,7 @@
 import composeClasses from '@mui/utils/composeClasses';
 import generateUtilityClass from '@mui/utils/generateUtilityClass';
 import generateUtilityClasses from '@mui/utils/generateUtilityClasses';
-import { createSlotArrayMap } from '../../internals/keyToArrayMap';
+import { createSlotArrayMap } from '@mui/x-internals/createSlotArrayMap';
 
 /**
  * @deprecated Use `RadarClasses` from `../radarClasses` instead.
