+++ packages/x-internals/src/createSlotArrayMap/index.ts
@@ -1,0 +1,1 @@
+export { createSlotArrayMap } from './createSlotArrayMap';
